refactor: language level to own traversal plugin
